+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - fix(uninstaller): include deletion of IdentityCenter and ResourcePolicyEnforcement stacks
+- fix(ssm): updated session manager role to allow kms permissions in all enabled regions
 
 ### Changed